Move FILAMENT_RUNOUT_SENSOR to a feature
